Did the stuff and stuff. Lot's of stuff nothing works anymore. Sadness

+++ SharedCore/graphics/textureatlas.d
@@ -6,7 +6,7 @@ import util.hash, collections.table, graphics.frame,
 struct SourceRect
 {
 	HashID hash;
-	float4 source;
+	short4 source;
 }
 
 struct TextureAtlas
